+++ services/dbt.go
@@ -10,7 +10,7 @@ import (
 	"path/filepath"
 	"time"
 
-	pbsql "github.com/streamingfast/substreams-sink-sql/pb/sf/substreams/sink/sql/v1"
+	pbsql "github.com/streamingfast/substreams-sink-sql/pb/sf/substreams/sink/sql/services/v1"
 	"go.uber.org/zap"
 )