+++ cmd/substreams-sink-sql/from_proto.go
@@ -17,6 +17,7 @@ import (
 	"github.com/streamingfast/substreams-sink-sql/db_proto/sql/postgres"
 	schema2 "github.com/streamingfast/substreams-sink-sql/db_proto/sql/schema"
 	stats2 "github.com/streamingfast/substreams-sink-sql/db_proto/stats"
+	pbsql "github.com/streamingfast/substreams-sink-sql/pb/sf/substreams/sink/sql/services/v1"
 	"github.com/streamingfast/substreams-sink-sql/services"
 	"github.com/streamingfast/substreams/manifest"
 	"go.uber.org/zap"
@@ -24,9 +25,9 @@ import (
 )
 
 var fromProtoCmd = Command(fromProtoE,
-	"from-proto <dsn> <manifest>",
+	"from-proto <dsn> <manifest> [output-module]",
 	"",
-	ExactArgs(2),
+	RangeArgs(2, 3),
 	Flags(func(flags *pflag.FlagSet) {
 		sink.AddFlagsToSet(flags, ignoreUndoBufferSize{})
 		flags.StringP("substreams-endpoint", "e", "", "Substreams gRPC endpoint. If empty, will be replaced by the SUBSTREAMS_ENDPOINT_{network_name} environment variable, where `network_name` is determined from the substreams manifest. Some network names have default endpoints.")
@@ -58,6 +59,11 @@ func fromProtoE(cmd *cobra.Command, args []string) error {
 
 	dsnString := args[0]
 	manifestPath := args[1]
+
+	outputModuleName := sink.InferOutputModuleFromPackage
+	if len(args) == 3 {
+		outputModuleName = args[2]
+	}
 
 	useConstraints := !sflags.MustGetBool(cmd, "no-constraints")
 	//useTransactions := !sflags.MustGetBool(cmd, "no-transactions")
@@ -110,12 +116,12 @@ func fromProtoE(cmd *cobra.Command, args []string) error {
 	}
 
 	//todo: handle params
-	spkg, module, _, _, err := sink.ReadManifestAndModuleAndBlockRange(manifestPath, "", nil, sink.InferOutputModuleFromPackage, "", false, "", zlog)
+	spkg, module, _, _, err := sink.ReadManifestAndModuleAndBlockRange(manifestPath, "", nil, outputModuleName, "", false, "", zlog)
 	if err != nil {
 		return fmt.Errorf("reading manifest: %w", err)
 	}
 
-	outputModuleName := module.Name
+	outputModuleName = module.Name
 	outputType := proto.ModuleOutputType(spkg, outputModuleName)
 	if outputType == "" {
 		return fmt.Errorf("could not find output type for module %s", outputModuleName)
@@ -123,7 +129,7 @@ func fromProtoE(cmd *cobra.Command, args []string) error {
 
 	service, err := extractSinkService(spkg)
 	if err != nil {
-		return fmt.Errorf("extracting sink service: %w", err)
+		service = &pbsql.Service{}
 	}
 
 	err = services.Run(service, zlog)