+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## v4.2.1
-Bump substreams to v1.10.3 to support new manifest data like `protobuf:excludePaths`
+
+* Bump substreams to v1.10.3 to support new manifest data like `protobuf:excludePaths`
 
 ## v4.2.0